+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixed locale cookie fallback [#221](https://github.com/cretueusebiu/laravel-vue-spa/pull/221)
 - Added scroll delay [#220](https://github.com/cretueusebiu/laravel-vue-spa/pull/220)
 - Added redirect to intended url [#208](https://github.com/cretueusebiu/laravel-vue-spa/pull/208)
+- Added middleware parameter [#124](https://github.com/cretueusebiu/laravel-vue-spa/pull/124)
 
 ## 4.7.0 - 2020-04-09